requirements.txt
# # YOLOv5 requirements
# # Usage: pip install -r requirements.txt
#
# # Base ------------------------------------------------------------------------
# gitpython>=3.1.30
# matplotlib>=3.3
# numpy>=1.22.2
# opencv-python>=4.1.1
# Pillow>=7.1.2
# psutil~=5.9.5
# PyYAML>=5.3.1
# requests>=2.23.0
# scipy>=1.4.1
# thop>=0.1.1  # FLOPs computation
# torch>=1.8.0  # see https://pytorch.org/get-started/locally (recommended)
# torchvision>=0.9.0
# tqdm>=4.64.0
# ultralytics>=8.0.147
# # protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012
#
# # Logging ---------------------------------------------------------------------
# # tensorboard>=2.4.1
# # clearml>=1.2.0
# # comet
#
# # Plotting --------------------------------------------------------------------
# pandas>=1.1.4
# seaborn>=0.11.0
#
# # Export ----------------------------------------------------------------------
# # coremltools>=6.0  # CoreML export
# # onnx>=1.10.0  # ONNX export
# # onnx-simplifier>=0.4.1  # ONNX simplifier
# # nvidia-pyindex  # TensorRT export
# # nvidia-tensorrt  # TensorRT export
# # scikit-learn<=1.1.2  # CoreML quantization
# # tensorflow>=2.4.0  # TF exports (-cpu, -aarch64, -macos)
# # tensorflowjs>=3.9.0  # TF.js export
# # openvino-dev>=2023.0  # OpenVINO export
#
# # Deploy ----------------------------------------------------------------------
# setuptools>=65.5.1 # Snyk vulnerability fix
# # tritonclient[all]~=2.24.0
#
# # Extras ----------------------------------------------------------------------
# # ipython  # interactive notebook
# # mss  # screenshots
# # albumentations>=1.0.3
# # pycocotools>=2.0.6  # COCO mAP


# Ultralytics requirements YOLO 8
# Example: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy==1.22.2 # pinned by Snyk to avoid a vulnerability
opencv-python>=4.6.0
pillow>=7.1.2
pyyaml>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.8.0
torchvision>=0.9.0
tqdm>=4.64.0

# Logging -------------------------------------
# tensorboard>=2.13.0
# dvclive>=2.12.0
# clearml
# comet

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=7.0.b1  # CoreML export
# onnx>=1.12.0  # ONNX export
# onnxsim>=0.4.1  # ONNX simplifier
# nvidia-pyindex  # TensorRT export
# nvidia-tensorrt  # TensorRT export
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TF exports (-cpu, -aarch64, -macos)
# tflite-support
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev>=2023.0  # OpenVINO export

# Extras --------------------------------------
psutil  # system utilization
py-cpuinfo  # display CPU info
# thop>=0.1.1  # FLOPs computation
# ipython  # interactive notebook
# albumentations>=1.0.3  # training augmentations
# pycocotools>=2.0.6  # COCO mAP
# roboflow

Flask~=2.3.2
imutils~=0.5.4